requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-ftp==3.6.0
apache-airflow-providers-http==4.6.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
appdirs==1.4.4
appnope==0.1.3
argcomplete==3.1.3
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backoff==1.10.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.3
bpython==0.24
cachelib==0.9.0
cachetools==5.3.2
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.0
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.2.0
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
curtsies==0.4.2
cwcwidth==0.1.9
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.8
Django==5.0
djangorestframework==3.14.0
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
executing==2.0.1
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.45.1
frozendict==2.3.9
frozenlist==1.4.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
google-re2==1.1
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.15.0
kiwisolver==1.4.5
kubernetes==28.1.0
lazy-object-proxy==1.9.0
libclang==16.0.6
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.2.0
mplfinance==0.12.10b0
multidict==6.0.4
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pathspec==0.11.2
peewee==3.17.0
pendulum==2.1.2
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
prison==0.2.1
prompt-toolkit==3.0.41
protobuf==4.24.4
psutil==5.9.6
psycopg==3.1.13
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.6.0
rich-argparse==1.4.0
rpds-py==0.10.6
rsa==4.9
seaborn==0.13.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.3.0
text-unidecode==1.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.18
virtualenv==20.25.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0
